On Windows, query console directly, not stdout.
On Windows, we must manually enable VT processing on the console. Prior to this
commit, we attempted to obtain a handle to the console by querying stdout. If
stdout has been redirected, however, then the handle doesn't correspond to the
console but instead to the redirected output. As a result, calls to enable VT
processing fail. By default, this results in `yansi` disabling coloring.

While this ordinarily doesn't have an impact on the expected functionality as
the console and stdout are one and the same, it disproportionately affects
running Rust unit tests that assume `yansi` coloring is in effect. This is
because Cargo captures stdout, redirecting it away from the console. This commit
resolves this issue, allowing unit tests to work as expected.